JZ47XX efuse driver.
The efuse driver give access to an 8KB OTP ROM in which can be read and
stored some important data.
It is controled by gpio 132 that belongs to gpiochip128.
By default CI20.dts put this gpio at high value, disabling the efuse,
but once this gpio is set at a low value it can be accessed.